requirements.txt
Django>=5.1,<6.0
fastapi>=0.116,<1.0
gunicorn>=23.0,<24.0
psycopg[binary]>=3.2,<4.0
SQLAlchemy>=2.0,<3.0
uvicorn[standard]>=0.35,<1.0
python-telegram-bot>=21.7,<22.0